Refactor configuration imports to use the new directory structure and add a readme for configuration clarification
