+++ src/api/configuration/readme.md
@@ -1,0 +1,5 @@
+
+#### What's the difference between `config` and `storage`?
+
+* **config** is the idea of user configurable configuration files. This is things like connection settings, actions, and the likes.
+* **storage** if for storing data that is not editable by the user. Details like history, authorisation lists, etc.
