+++ src/api/tests/globalSetup.ts
@@ -3,8 +3,8 @@ import IBMi from "../IBMi";
 import { ENV_CREDS } from "./env";
 import { getConnection, setConnection } from "./state";
 import { afterAll, beforeAll, expect } from "vitest";
-import { CodeForIStorage, ConnectionStorage, VirtualStorage } from "../configuration/Storage";
-import { VirtualConfig } from "../configuration/Config";
+import { CodeForIStorage, ConnectionStorage, VirtualStorage } from "../configuration/storage/CodeForIStorage";
+import { VirtualConfig } from "../configuration/config/VirtualConfig";
 
 beforeAll(async () => {
   const virtualStorage = new VirtualStorage();
