+++ src/ui/views/ConnectionBrowser.ts
@@ -4,7 +4,7 @@ import { ConnectionData, Server } from '../../typings';
 import { instance } from '../../instantiate';
 import { Login } from '../../webviews/login';
 import IBMi from '../../api/IBMi';
-import { ConnectionConfig, ConnectionManager } from '../../api/configuration/ConnectionManager';
+import { ConnectionConfig, ConnectionManager } from '../../api/configuration/config/ConnectionManager';
 import { deleteStoredPassword, getStoredPassword, setStoredPassword } from '../../config/passwords';
 
 type CopyOperationItem = {
